requirements.txt
asgiref==3.4.1
attrs==22.2.0
black==22.8.0
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.0.4
click-plugins==1.1.1
cligj==0.7.2
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.8
Django==3.2.18
django-debug-toolbar==3.2.4
django-filter==21.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==4.4.0
drf-spectacular==0.26.0
drf-yasg==1.21.5
Fiona==1.8.22
geopandas==0.9.0
idna==3.4
importlib-metadata==4.8.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==3.2.0
MarkupSafe==2.0.1
munch==2.5.0
mypy-extensions==1.0.0
numpy==1.19.5
openapi-codec==1.3.2
packaging==21.3
pandas==1.1.5
pathspec==0.9.0
platformdirs==2.4.0
psycopg2-binary==2.9.5
PyJWT==1.7.1
pyparsing==3.0.9
pyproj==3.0.1
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
Shapely==1.8.5.post1
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.3
tomli==1.2.3
typed-ast==1.5.4
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.14
zipp==3.6.0
